01-nlp-synthetic-languages/master-data-science-portfolio/02-real-time-fraud-detection/producer/requirements.txt
kafka-python==2.2.14
faker==22.0.0